Added print email into email_obj.py and deleted print_email.py
